refactor: improving CC, Hasteald and cohesion metrics for the codebase
